Use Qt type typedefs
As for instance char might be an unsigned char on Arm.

Task-Id: KUE-580

+++ src/core/gltf2importer/bufferaccessorparser.cpp
@@ -65,7 +65,7 @@ QVector<float> jsonArrayToVectorOfFloats(const QJsonArray &values)
     return d;
 }
 
-quint32 accessorDataSizeFromJson(const QString &type)
+quint8 accessorDataSizeFromJson(const QString &type)
 {
     const QString typeName = type.toUpper();
     if (typeName == QLatin1String("SCALAR"))
